requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic-settings>=2.5.0
httpx>=0.27.0
sqlalchemy>=2.0.35
pymysql>=1.1.0
cryptography>=43.0.0
chromadb>=0.5.23
elasticsearch>=8.14.0,<9.0.0
python-docx>=1.1.2
python-multipart>=0.0.9
